pyproject.toml
[tool.poetry]
name = "sites-stat"
version = "0.1.0"
description = ""
authors = ["linar <chevianchik@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
uWSGI = "2.0.21"
django = ">=4.1.2,<4.2.0"
psycopg2-binary = ">=2.9.3,<2.10.0"
infi-clickhouse-orm = "^2.1.3"
djangorestframework = "^3.14.0"
httpagentparser = "^1.9.5"
django-webpack-loader = "^1.8.1"
django-cors-headers = "^3.14.0"
celery = "^5.2.7"
redis = "^4.5.4"
django-celery-beat = "^2.5.0"
pre-commit = "^3.2.2" # TODO перенести в dev dependencies
virtualenv = "^20.21.0"



[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
